+++ ledger/sandbox/src/main/scala/com/digitalasset/platform/resources/ResourceOwner.scala
@@ -8,3 +8,8 @@ import scala.concurrent.ExecutionContext
 trait ResourceOwner[T] {
   def acquire()(implicit _executionContext: ExecutionContext): Resource[T]
 }
+
+object ResourceOwner {
+  def apply[T <: AutoCloseable](acquireCloseable: () => T): ResourceOwner[T] =
+    new AutoCloseableResourceOwner[T](acquireCloseable)
+}
